local-cloud/lambdas/websocket-authorizer/requirements.txt
PyJWT==2.8.0
boto3==1.29.0
cryptography==41.0.7